service/requirements.txt
functions-framework==3.0.0
flask==2.0.3
google-cloud-error-reporting==1.5.2